WIP: change to 16 bit floats (except conv math for now)

+++ src/gi/shaders/rp_main_descriptors.glsl
@@ -70,7 +70,7 @@ layout(binding = BINDING_INDEX_AOV_INSTANCE_ID, std430) writeonly buffer Instanc
 layout(binding = BINDING_INDEX_AOV_DOUBLE_SIDED, std430) writeonly buffer DoubleSidedBuffer { vec3 DoubleSidedAov[]; };
 #endif
 #if (AOV_MASK & AOV_BIT_OIDN) != 0
-layout(binding = BINDING_INDEX_AOV_OIDN, std430) writeonly buffer OidnBuffer { float OidnAov[]; }; // 9 channels: color, normal, albedo
+layout(binding = BINDING_INDEX_AOV_OIDN, std430) writeonly buffer OidnBuffer { float16_t OidnAov[]; }; // 9 channels: color, normal, albedo
 #endif
 
 layout(set = 1, binding = BINDING_INDEX_TEXTURES_2D) uniform texture2D textures_2d[MAX_TEXTURE_COUNT];
